Update Functions.py
Added the log functions. Entering 'exit' when asked for a move will save a log of the match and close the game.
Entering 'debug' when asked for a move will print a log of the match and ask you if you'd like to save the log.
You'll also be asked if you'd like to save a log after the match.